+++ changelog.md
@@ -1,4 +1,4 @@
-### ?, ? - V1.4.1.
+### March 24, 2024 - V1.4.1.
 
 - Add some color schemes.
 - Add version info to gui.